Max threads, timeout added, documentation updated.
